fix: fix casting  - fix returning query

+++ service/bulk/schema/750$bulk.batch.revertStatus.sql
@@ -37,7 +37,7 @@ BEGIN
                     bh."batchHistoryId" DESC
                 LIMIT 
                     1
-            )::integer,
+            ),
             NULL,
             NULL,
             "@actorId",
